move(step2): package 구조 분리 (lotto, calculator, common)
