+++ src/main/java/common/ErrorMessage.java
@@ -1,4 +1,4 @@
-package util;
+package common;
 
 import calculator.Operator;
